feat: add book class contollers

+++ api/delete.php
@@ -5,21 +5,39 @@ header("Access-Control-Allow-Methods: POST");
 header("Access-Control-Max-Age: 3600");
 header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
 
-include_once '../config/database.php';
-include_once '../class/authors.php';
+include_once dirname(__DIR__) . '/config/database.php';
+include_once dirname(__DIR__) . '/class/authors.php';
+include_once dirname(__DIR__) . '/class/books.php';
 
-$database = new Database();
-$db = $database->getConnection();
+function getData($class) {
+    $database = new Database();
+    $db = $database->getConnection();
+    $data = json_decode(file_get_contents("php://input"));
 
-$item = new Authors($db);
+    if($class == 'authors') {
+        $item = new Authors($db);
+        $item->id = $data->id;
 
-$data = json_decode(file_get_contents("php://input"));
+        if($item->deleteAuthor()){
+            echo json_encode("Author deleted.");
+        } else{
+            echo json_encode("Author could not be deleted.");
+        }
+    }
+    if($class == 'books') {
+        $item = new Books($db);
+        $item->id = $data->id;
 
-$item->id = $data->id;
+        if($item->deleteBook()){
+            echo json_encode("Book deleted.");
+        } else{
+            echo json_encode("Book could not be deleted.");
+        }
+    }
 
-if($item->deleteAuthor()){
-    echo json_encode("Author deleted.");
-} else{
-    echo json_encode("Data could not be deleted.");
+
+
+
 }
+
 ?>
